Move Completions.scala to completions/Completions.scala
